requirements.txt
Django
Pillow==2.5.3
Unidecode==0.04.16
beautifulsoup4==4.3.2
django-appconf==0.6
django-compressor==1.4
django-libsass==0.2
django-modelcluster==0.4
django-taggit==0.12.1
django-treebeard==2.0
html5lib==0.999
libsass==0.5.0
requests==2.3.0
six==1.7.3
unicodecsv==0.9.4
wagtail==0.5
wsgiref==0.1.2
